Save coverage for each page in a test
Save coverage for each page visited in a test
by keeping a reference to the coverage for each app window loaded
and then saving the coverage at the end after coverage is calculated.

An additional page is added to the example test app
and the page test navigates to the new page.
Without this coverage change,
coverage is lost for `app.js` due to the last page not loading it.
With this coverage change,
coverage is kept for `app.js` along with coverage for `about.js`.